+++ pkg/azurecluster/validate_update_test.go
@@ -18,7 +18,6 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 		name            string
 		oldAzureCluster []byte
 		newAzureCluster []byte
-		allowed         bool
 		errorMatcher    func(err error) bool
 	}
 
@@ -27,21 +26,18 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 			name:            "case 0: unchanged ControlPlaneEndpoint",
 			oldAzureCluster: azureClusterRawObject("ab123", "api.ab123.test.westeurope.azure.gigantic.io", 443),
 			newAzureCluster: azureClusterRawObject("ab123", "api.ab123.test.westeurope.azure.gigantic.io", 443),
-			allowed:         true,
 			errorMatcher:    nil,
 		},
 		{
 			name:            "case 1: host changed",
 			oldAzureCluster: azureClusterRawObject("ab123", "api.ab123.test.westeurope.azure.gigantic.io", 443),
 			newAzureCluster: azureClusterRawObject("ab123", "api.azure.gigantic.io", 443),
-			allowed:         false,
 			errorMatcher:    errors.IsInvalidOperationError,
 		},
 		{
 			name:            "case 2: port changed",
 			oldAzureCluster: azureClusterRawObject("ab123", "api.ab123.test.westeurope.azure.gigantic.io", 443),
 			newAzureCluster: azureClusterRawObject("ab123", "api.ab123.test.westeurope.azure.gigantic.io", 80),
-			allowed:         false,
 			errorMatcher:    errors.IsInvalidOperationError,
 		},
 	}
@@ -64,7 +60,7 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 			}
 
 			// Run admission request to validate AzureConfig updates.
-			allowed, err := admit.Validate(context.Background(), getUpdateAdmissionRequest(tc.oldAzureCluster, tc.newAzureCluster))
+			err = admit.Validate(context.Background(), getUpdateAdmissionRequest(tc.oldAzureCluster, tc.newAzureCluster))
 
 			// Check if the error is the expected one.
 			switch {
@@ -76,11 +72,6 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 				t.Fatalf("expected %#v got %#v", "error", nil)
 			case !tc.errorMatcher(err):
 				t.Fatalf("unexpected error: %#v", err)
-			}
-
-			// Check if the validation result is the expected one.
-			if tc.allowed != allowed {
-				t.Fatalf("expected %v to be equal to %v", tc.allowed, allowed)
 			}
 		})
 	}