+++ pkg/machinepool/validate_machinepool_create.go
@@ -47,23 +47,23 @@ func NewCreateValidator(config CreateValidatorConfig) (*CreateValidator, error) 
 	return admitter, nil
 }
 
-func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) (bool, error) {
+func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) error {
 	machinePoolNewCR := &v1alpha3.MachinePool{}
 	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, machinePoolNewCR); err != nil {
-		return false, microerror.Maskf(parsingFailedError, "unable to parse machinePool CR: %v", err)
+		return microerror.Maskf(parsingFailedError, "unable to parse machinePool CR: %v", err)
 	}
 
 	err := generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, machinePoolNewCR)
 	if err != nil {
-		return false, microerror.Mask(err)
+		return microerror.Mask(err)
 	}
 
 	err = a.checkAvailabilityZones(ctx, machinePoolNewCR)
 	if err != nil {
-		return false, microerror.Mask(err)
+		return microerror.Mask(err)
 	}
 
-	return true, nil
+	return nil
 }
 
 func (a *CreateValidator) Log(keyVals ...interface{}) {